requirements.txt
-i https://pypi.org/simple
amqp==2.4.0
billiard==3.5.0.5
celery==4.2.1
dj-database-url==0.5.0
django-celery-results==1.0.4
django-crontab==0.7.1
django-extensions==2.1.4
django-redis==4.10.0
django==2.1.5
djangorestframework-jwt==1.11.0
djangorestframework==3.9.1
kombu==4.2.2.post1
pillow==5.4.1
psycopg2-binary==2.7.6.1
pyjwt==1.7.1
python-dotenv==0.10.1
pytz==2018.9
redis==3.0.1
six==1.12.0
vine==1.2.0
